Made signin and login disappear when the other is clicked
